refactor(gatekeep): modularize security check and validation scripts
