+++ .gitignore
@@ -80,6 +80,7 @@ vox
 # Specific Scripts & Ignores
 # ==============================================================================
 ai_gatekeeper.log
+codeflow
 
 # ==============================================================================
 # Arquivos do PyTest
